LPD-73835 SF - Move to d.d.m.util package
